update: rename component source files for matching its actual name
