--- src/views/components/SearchForm/ISearchForm.ts
@@ -1,4 +1,0 @@
-import { ComponentProps } from 'react'
-
-
-export interface ISearchForm extends ComponentProps<'form'> {}
