+++ src/views/components/SearchWidget/ISearchWidget.ts
@@ -1,0 +1,4 @@
+import { ComponentProps } from 'react'
+
+
+export interface ISearchWidget extends ComponentProps<'form'> {}
