+++ src/views/components/SearchWidget/SearchWidget.module.scss
