+++ src/views/components/index.ts
@@ -1,4 +1,4 @@
 export * from './Footer/Footer'
 export * from './Header/Header'
-export * from './SearchForm/SearchWidget.tsx'
+export * from '@views/components/SearchWidget/SearchWidget.tsx'
 export * from './GameCard/GameCard'
